requirements.txt
docopt==0.6.2
python-qpid-proton==0.24.0
stomp.py==4.1.21
